test_gemini_1_5_flash.py
#!/usr/bin/env python3
"""
Test that gemini-1.5-flash works with the updated Beige AI system
"""
import google.generativeai as genai
import sys
import os

print("Testing Gemini 1.5 Flash Model with System Instruction...")
print("-" * 60)

# Get API key from environment or use hardcoded test key
api_key = os.getenv('GEMINI_API_KEY', 'AIzaSyBb_2ZouURAbi-fBR8jL_uwEGthGSlBu2E')

# Configure API key
genai.configure(api_key=api_key)

# Test 1: Basic model initialization
print("\n✓ TEST 1: Initialize genai.GenerativeModel('gemini-1.5-flash')")
try:
    model = genai.GenerativeModel('gemini-1.5-flash')
    print("  ✅ Model initialized successfully")
except Exception as e:
    print(f"  ❌ FAILED: {e}")
    sys.exit(1)

# Test 2: Model with system prompt in content
print("\n✓ TEST 2: Initialize model and include system prompt in content")
try:
    system_prompt = "You are a helpful assistant. Always respond concisely in 1-2 sentences."
    model = genai.GenerativeModel('gemini-1.5-flash')
    prompt_with_system = f"{system_prompt}\n\nRespond to this: Say hello"
    response = model.generate_content(prompt_with_system, generation_config={'max_output_tokens': 20})
    if response and response.text:
        print(f"  ✅ Model created and system prompt works")
    else:
        print(f"  ❌ Empty response")
        sys.exit(1)
except Exception as e:
    print(f"  ❌ FAILED: {e}")
    sys.exit(1)

# Test 3: Generate content with configuration
print("\n✓ TEST 3: Call generate_content() with generation_config")
try:
    response = model.generate_content(
        "Confirm Beige AI Concierge works with gemini-1.5-flash",
        generation_config={
            'temperature': 0.8,
            'top_p': 0.95,
            'top_k': 40,
            'max_output_tokens': 150
        }
    )
    if response and response.text:
        print(f"  ✅ Response: {response.text[:100]}")
    else:
        print("  ❌ Empty response")
        sys.exit(1)
except Exception as e:
    print(f"  ❌ FAILED: {e}")
    sys.exit(1)

# Test 4: Verify Concierge-style system prompt works
print("\n✓ TEST 4: Test with actual Concierge system prompt")
try:
    from backend.concierge_system_prompt import CONCIERGE_SYSTEM_PROMPT
    
    model_with_concierge = genai.GenerativeModel('gemini-1.5-flash')
    
    test_prompt_content = """User context:
- Current mood: happy
- Weather: sunny
- Time of day: afternoon
- Sweetness preference: 5/10
- Health consciousness: 8/10

Recommended cake:
- Name: Berry Garden Cake
- Flavor profile: fresh, fruity
- Category: light & refreshing

Generate a personalized recommendation..."""
    
    # Include system prompt in content
    full_prompt = f"{CONCIERGE_SYSTEM_PROMPT}\n\n{test_prompt_content}"
    
    response = model_with_concierge.generate_content(
        full_prompt,
        generation_config={'temperature': 0.8, 'max_output_tokens': 150}
    )
    
    if response and response.text:
        output = response.text.strip()
        print(f"  ✅ Concierge response generated")
        print(f"  Sample output: {output[:150]}...")
        
        # Verify it doesn't contain robotic markers
        bad_markers = ['confidence', '%', 'score', 'probability']
        has_bad_markers = any(marker.lower() in output.lower() for marker in bad_markers)
        if not has_bad_markers:
            print(f"  ✅ Output is properly editorial (no data exposure)")
        else:
            print(f"  ⚠️  Output contains technical markers")
    else:
        print("  ❌ Empty response")
        sys.exit(1)
        
except ImportError:
    print("  ⚠️  SKIPPED (concierge_system_prompt not available)")
except Exception as e:
    print(f"  ❌ FAILED: {e}")
    sys.exit(1)

print("\n" + "=" * 60)
print("✅ ALL TESTS PASSED")
print("Gemini 1.5 Flash is fully integrated and working correctly")
print("=" * 60)
