requirements.txt
streamlit>=1.31.0
pandas>=2.0.3
numpy>=2.0.0
scikit-learn>=1.8.0
matplotlib>=3.7.2
joblib>=1.3.1
requests>=2.31.0
google-generativeai>=0.8.0
protobuf>=3.20.0
